Move StructuredOutputSerializer to feature namespace
