+++ src/platform/src/StructuredOutput/Serializer.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\AI\Platform\Serializer;
+namespace Symfony\AI\Platform\StructuredOutput;
 
 use Symfony\Component\PropertyInfo\Extractor\PhpDocExtractor;
 use Symfony\Component\PropertyInfo\Extractor\ReflectionExtractor;
@@ -21,9 +21,9 @@ use Symfony\Component\Serializer\Mapping\Loader\AttributeLoader;
 use Symfony\Component\Serializer\Normalizer\ArrayDenormalizer;
 use Symfony\Component\Serializer\Normalizer\BackedEnumNormalizer;
 use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
-use Symfony\Component\Serializer\Serializer;
+use Symfony\Component\Serializer\Serializer as BaseSerializer;
 
-class StructuredOutputSerializer extends Serializer
+class Serializer extends BaseSerializer
 {
     /*
      * Custom serializer made to deserialize StructuredOutput.
